+++ tests/keyfiles.rs
@@ -7,7 +7,7 @@ use std::str::from_utf8;
 
 mod utils;
 
-const TEST_FILE_PASS: &[u8] = b"12345678";
+const TEST_FILE_PASS: &str = "12345678";
 
 fn locate_crate_files<P: AsRef<Path>>(path: P) -> PathBuf {
     let mut abspath = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
@@ -15,7 +15,7 @@ fn locate_crate_files<P: AsRef<Path>>(path: P) -> PathBuf {
     abspath
 }
 
-fn verify_key<P: AsRef<Path>>(keyfile: P, passphrase: Option<&[u8]>) {
+fn verify_key<P: AsRef<Path>>(keyfile: P, passphrase: Option<&str>) {
     let keypath = locate_crate_files(keyfile);
     let pubkeypath = keypath.with_extension("pub");
 
@@ -112,23 +112,23 @@ fn keyfile_pkcs8_rsa_enc() {
 #[test]
 #[should_panic]
 fn keyfile_pem_rsa_wrong() {
-    verify_key("assets/pem_rsa_enc", Some(b"deadbeef"));
+    verify_key("assets/pem_rsa_enc", Some("deadbeef"));
 }
 
 #[test]
 #[should_panic]
 fn keyfile_pem_dsa_wrong() {
-    verify_key("assets/pem_dsa_enc", Some(b"hashdjf"));
+    verify_key("assets/pem_dsa_enc", Some("hashdjf"));
 }
 
 #[test]
 #[should_panic]
 fn keyfile_pem_ecdsa_wrong() {
-    verify_key("assets/pem_ecdsa_enc", Some(b"1234567"));
+    verify_key("assets/pem_ecdsa_enc", Some("1234567"));
 }
 
 #[test]
 #[should_panic]
 fn keyfile_pem_ed25519_wrong() {
-    verify_key("assets/pem_ed25519_enc", Some(b"^&@#Y&G*"));
+    verify_key("assets/pem_ed25519_enc", Some("^&@#Y&G*"));
 }